requirements.txt
GEOparse
cptac~=1.5.14
biopython
scikit-learn
scipy~=1.14.1
numpy~=2.0.2
pandas~=2.2.2
sqlalchemy~=2.0.36
pymongo~=4.10.1
rpy2
psycopg2
typing~=3.7.4.3
pytest~=8.3.2
tqdm
python-dotenv~=1.0.1
requests~=2.32.3
alembic~=1.14.0
retry
lxml~=5.3.0
pydantic
matplotlib~=3.9.2
seaborn~=0.13.2
gprofiler-official
scanpy
psutil~=6.0.0
pathlib~=1.0.1
psutil~=6.0.0
statsmodels~=0.14.4
graphviz~=0.20.3

argparse~=1.4.0
PyYAML~=6.0.2